Move visualizeTree to RRT package
